Split pane drop routing session types for #7529
